Reorganise Server tests
