+++ server/DataDeliveryTrigger/index.ts
@@ -3,7 +3,7 @@ import {EnvironmentVariables} from "../Config";
 import * as PinoHttp from "pino-http";
 import {SendAPIRequest} from "../SendRequest";
 
-export default function DataDelivery(environmentVariables: EnvironmentVariables, logger: PinoHttp.HttpLogger): Router {
+export default function DataDeliveryTrigger(environmentVariables: EnvironmentVariables, logger: PinoHttp.HttpLogger): Router {
     const {PROJECT_ID, AZURE_AUTH_TOKEN, ENV_NAME, GIT_BRANCH, DATA_DELIVERY_AZURE_PIPELINE_NO}: EnvironmentVariables = environmentVariables;
     const router = express.Router();
